Improve systemd detection for diagnostics.
Previous code assumed that if systemctl exists, we can check unit status.
However in a containerized deployment systemctl may be in the container but
non-functional, resulting in multiple false errors checking unit status.

Fixed by checking that systemctl is actually usable before we proceed with
checking unit status.

Additionally added a check that journalctl is usable in the AnalyzeLogs
CanRun() method, resulting in slightly more accurate output in a container
indicating that the check was skipped entirely.